+++ routes/api/protected/user.php
@@ -8,12 +8,16 @@ use Wave8\Factotum\Base\Models\User;
 Route::prefix('users')
     ->controller(UserController::class)
     ->group(function () {
-        Route::get('', 'index')->can('read', User::class);
+        Route::get('', 'index')->can('filter', User::class);
         Route::get('{user}', 'show')->can('read', 'user');
-        Route::post('', 'store')->can('create', User::class);
-        Route::put('{user}', 'update')->can('update', 'user');
-        Route::delete('{user}', 'destroy')->can('delete', 'user');
-
-        Route::put('{id}/settings/{setting}', 'updateSetting')->can('updateUserSetting', 'setting');
         Route::get('{id}/settings', 'settings')->can('read', Setting::class);
+
+        Route::post('', 'store')->can('create', User::class);
+
+        Route::put('{user}', 'update')->can('update', 'user');
+        Route::put('{id}/settings/{setting}', 'updateSetting')->can('updateUserSetting', 'setting');
+
+        Route::patch('{user}/change-password', 'changePassword')->can('changePassword', 'user');
+
+        Route::delete('{user}', 'destroy')->can('delete', 'user');
     });