+++ src/Models/PasswordHistory.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Wave8\Factotum\Base\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+
+class PasswordHistory extends Model
+{
+    use HasFactory;
+
+    protected $fillable = [
+        'user_id',
+        'password',
+        'expires_at',
+    ];
+
+    protected $hidden = [
+        'password',
+    ];
+
+    protected $casts = [
+        'expires_at' => 'datetime',
+        'password' => 'hashed',
+    ];
+
+    public function user(): BelongsTo
+    {
+        return $this->belongsTo(User::class);
+    }
+}