+++ src/Services/Api/UserService.php
@@ -14,9 +14,18 @@ class UserService implements UserServiceInterface
 
     public function create(Data $data): Model
     {
-        return $this->user::create(
+        $user = $this->user::create(
             attributes: $data->toArray()
         );
+
+        if ($data->password) {
+            $user->password_histories()->create([
+                'password' => $user->password,
+                'expires_at' => now()->addDays(config('factotum_base.auth.password_expiration_days')),
+            ]);
+        }
+
+        return $user;
     }
 
     public function read(int $id): Model
@@ -31,6 +40,21 @@ class UserService implements UserServiceInterface
         $user->update(
             attributes: $data->toArray()
         );
+
+        return $user;
+    }
+
+    public function updatePassword(User $user, string $password): User
+    {
+        $user = $this->user::findOrFail($user->id);
+
+        $user->password = $password;
+        $user->save();
+
+        $user->password_histories()->create([
+            'password' => $user->password,
+            'expires_at' => now()->addDays(config('factotum_base.auth.password_expiration_days')),
+        ]);
 
         return $user;
     }